исправления по замечаниям
