+++ src/login.js
@@ -1,10 +1,5 @@
 import React from 'react';
 
-class Login extends React.Component {
-
-    render() {
-        return <h1>Login</h1>
-    }
-}
+const Login = () => <h1>Login</h1>;
 
 export default Login;
