+++ src/map.js
@@ -1,10 +1,5 @@
 import React from 'react';
 
-class Map extends React.Component {
-
-    render() {
-        return <h1>Map</h1>
-    }
-}
+const Map = () => <h1>Map</h1>;
 
 export default Map;
