+++ src/profile.js
@@ -1,10 +1,5 @@
 import React from 'react';
 
-class Profile extends React.Component {
-
-    render() {
-        return <h1>Profile</h1>
-    }
-}
+const Profile = () => <h1>Profile</h1>;
 
 export default Profile;
