+++ src/signup.js
@@ -1,10 +1,5 @@
 import React from 'react';
 
-class Signup extends React.Component {
-
-    render() {
-        return <h1>Signup</h1>
-    }
-}
+const Signup = () => <h1>Signup</h1>;
 
 export default Signup;
